Restructure query suggestions service
