+++ stroom-meta/stroom-meta-impl/build.gradle
@@ -4,7 +4,6 @@ dependencies {
     implementation project(':stroom-cluster:stroom-cluster-lock-api')
     implementation project(':stroom-cluster:stroom-cluster-lock-mock')
     implementation project(':stroom-config:stroom-config-common')
-    implementation project(':stroom-core')
     implementation project(':stroom-core-shared')
     implementation project(':stroom-data:stroom-data-retention-api')
     implementation project(':stroom-docref')
@@ -22,6 +21,7 @@ dependencies {
     implementation project(':stroom-query:stroom-expression')
     implementation project(':stroom-query:stroom-query-api')
     implementation project(':stroom-query:stroom-query-common')
+    implementation project(':stroom-query:stroom-query-impl')
     implementation project(':stroom-search:stroom-expression-matcher')
     implementation project(':stroom-search:stroom-searchable-api')
     implementation project(':stroom-security:stroom-security-api')
