+++ stroom-meta/stroom-meta-impl/src/main/java/stroom/meta/impl/MetaModule.java
@@ -1,6 +1,5 @@
 package stroom.meta.impl;
 
-import stroom.core.query.SuggestionsServiceBinder;
 import stroom.event.logging.api.ObjectInfoProviderBinder;
 import stroom.job.api.ScheduledJobsBinder;
 import stroom.meta.api.MetaSecurityFilter;
@@ -8,6 +7,7 @@ import stroom.meta.api.MetaService;
 import stroom.meta.api.PhysicalDelete;
 import stroom.meta.shared.Meta;
 import stroom.meta.shared.MetaFields;
+import stroom.query.impl.SuggestionsServiceBinder;
 import stroom.searchable.api.Searchable;
 import stroom.util.RunnableWrapper;
 import stroom.util.guice.GuiceUtil;
