+++ stroom-meta/stroom-meta-impl/src/test/java/stroom.meta.impl/TestMetaSuggestionsQueryHandler.java
@@ -1,12 +1,12 @@
 package stroom.meta.impl;
 
-import stroom.core.query.SuggestionsService;
 import stroom.docref.DocRef;
 import stroom.docrefinfo.api.DocRefInfoService;
 import stroom.feed.api.FeedStore;
 import stroom.meta.api.MetaService;
 import stroom.meta.shared.MetaFields;
 import stroom.pipeline.PipelineStore;
+import stroom.query.impl.SuggestionsService;
 import stroom.query.shared.FetchSuggestionsRequest;
 import stroom.security.api.SecurityContext;
 import stroom.security.mock.MockSecurityContext;
