+++ stroom-query/stroom-query-impl/src/main/java/stroom/query/impl/SuggestionsService.java
@@ -1,4 +1,4 @@
-package stroom.core.query;
+package stroom.query.impl;
 
 import stroom.query.shared.FetchSuggestionsRequest;
