+++ stroom-query/stroom-query-impl/src/main/java/stroom/query/impl/SuggestionsServiceImpl.java
@@ -1,4 +1,4 @@
-package stroom.core.query;
+package stroom.query.impl;
 
 import stroom.query.shared.FetchSuggestionsRequest;
 import stroom.security.api.SecurityContext;
