+++ stroom-search/stroom-search-elastic/src/main/java/stroom/search/elastic/suggest/ElasticSuggestionsQueryHandler.java
@@ -1,6 +1,6 @@
 package stroom.search.elastic.suggest;
 
-import stroom.core.query.SuggestionsQueryHandler;
+import stroom.query.impl.SuggestionsQueryHandler;
 
 public interface ElasticSuggestionsQueryHandler extends SuggestionsQueryHandler {
